# docs/swagger.yml
basePath: /
consumes:
    - application/json
definitions:
    EventRequest:
        properties:
            end_date:
                format: date-time
                type: string
                x-go-name: EndDate
            event_name:
                type: string
                x-go-name: EventName
            start_date:
                format: date-time
                type: string
                x-go-name: StartDate
            status:
                format: int64
                type: integer
                x-go-name: Status
        type: object
        x-go-package: bigmind/xcheck-be/internal/models
    MetaResponse:
        properties:
            limit:
                format: int64
                type: integer
                x-go-name: Limit
            page:
                format: int64
                type: integer
                x-go-name: Page
            total:
                format: int64
                type: integer
                x-go-name: Total
        type: object
        x-go-package: bigmind/xcheck-be/utils
    RefreshToken:
        properties:
            refresh_token:
                type: string
                x-go-name: RefreshToken
        type: object
        x-go-package: bigmind/xcheck-be/internal/models
    UserLogin:
        properties:
            password:
                type: string
                x-go-name: Password
            username:
                type: string
                x-go-name: Username
        type: object
        x-go-package: bigmind/xcheck-be/internal/models
    UserRequest:
        properties:
            email:
                type: string
                x-go-name: Email
            password:
                type: string
                x-go-name: Password
            role_id:
                format: int64
                type: integer
                x-go-name: RoleID
            username:
                type: string
                x-go-name: Username
        type: object
        x-go-package: bigmind/xcheck-be/internal/models
info:
    title: 'XCheck API:'
    version: "0.1"
paths:
    /auth/me:
        get:
            description: Get current user
            operationId: authMe
            responses:
                "200":
                    description: ""
            security:
                - Bearer:
                    - '[]'
            tags:
                - Auth
    /auth/signin:
        post:
            description: Signin
            operationId: authSignin
            parameters:
                - in: body
                  name: UserLogin
                  required: true
                  schema:
                    $ref: '#/definitions/UserLogin'
            responses:
                "200":
                    description: ""
            tags:
                - Auth
    /auth/signout:
        post:
            description: Signout
            operationId: signout
            responses:
                "200":
                    description: ""
            security:
                - Bearer:
                    - '[]'
            tags:
                - Auth
    /auth/token:
        post:
            description: Refresh token
            operationId: authRefreshToken
            parameters:
                - in: body
                  name: RefreshToken
                  required: true
                  schema:
                    $ref: '#/definitions/RefreshToken'
                  x-go-name: UserLogin
            responses:
                "200":
                    description: ""
            security:
                - Bearer:
                    - '[]'
            tags:
                - Auth
    /events:
        get:
            description: Get Event list
            operationId: getEventList
            parameters:
                - format: int64
                  in: query
                  name: limit
                  required: true
                  type: integer
                  x-go-name: Limit
                - format: int64
                  in: query
                  name: page
                  required: true
                  type: integer
                  x-go-name: Page
            responses:
                "200":
                    description: ""
            security:
                - Bearer:
                    - '[]'
            tags:
                - Event
        post:
            description: Create Event
            operationId: createEvent
            parameters:
                - in: body
                  name: EventRequest
                  required: true
                  schema:
                    $ref: '#/definitions/EventRequest'
            responses:
                "200":
                    description: ""
            security:
                - Bearer:
                    - '[]'
            tags:
                - Event
    /events/{id}:
        delete:
            description: Delete Event by id
            operationId: deleteEvent
            parameters:
                - format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    description: ""
            security:
                - Bearer:
                    - '[]'
            tags:
                - Event
        get:
            description: Get Event by id
            operationId: getEvent
            parameters:
                - format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    description: ""
            security:
                - Bearer:
                    - '[]'
            tags:
                - Event
    /users:
        get:
            description: Get User list
            operationId: getUserList
            parameters:
                - format: int64
                  in: query
                  name: limit
                  required: true
                  type: integer
                  x-go-name: Limit
                - format: int64
                  in: query
                  name: page
                  required: true
                  type: integer
                  x-go-name: Page
            responses:
                "200":
                    description: ""
            security:
                - Bearer:
                    - '[]'
            tags:
                - User
        post:
            description: Create User
            operationId: createUser
            parameters:
                - in: body
                  name: UserRequest
                  required: true
                  schema:
                    $ref: '#/definitions/UserRequest'
            responses:
                "200":
                    description: ""
            security:
                - Bearer:
                    - '[]'
            tags:
                - User
    /users/{id}:
        get:
            description: Get User by id
            operationId: getUser
            parameters:
                - format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    description: ""
            security:
                - Bearer:
                    - '[]'
            tags:
                - User
produces:
    - application/json
responses:
    APIResponse:
        description: ""
        headers:
            code:
                format: int64
                type: integer
            data: {}
            message:
                type: string
            meta: {}
            refresh_token:
                type: string
            response_time:
                format: double
                type: number
            status:
                type: string
            token:
                type: string
        schema:
            $ref: '#/definitions/MetaResponse'
schemes:
    - http
    - https
securityDefinitions:
    Bearer:
        description: Bearer "token"
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"